+++ tests/testthat/test-inputs.R
@@ -32,7 +32,7 @@ test_that("pagfl inputs", {
   # Plain PGMM
   expect_no_error(pagfl(y ~ a + b, data = data, index = c("i", "t"), method = "PGMM", lambda = 1, Z = X, verbose = F))
   # Plain PGMM and bias correction
-  expect_no_error(pagfl(y ~ a + b, data = data, index = c("i", "t"), method = "PGMM", lambda = 1, Z = X, bias_correc = T))
+  expect_no_error(pagfl(y ~ a + b, data = data, index = c("i", "t"), method = "PGMM", lambda = 1, Z = X, bias_correc = T, verbose = F))
   # No method
   expect_error(pagfl(y ~ a + b, data = data, index = c("i", "t"), method = "A", lambda = 1, verbose = F))
   # Incorrect argument
@@ -45,20 +45,20 @@ test_that("pagfl inputs", {
   # No dependent variable
   expect_error(pagfl(~., data = data, index = c("i", "t"), lambda = 1, verbose = F))
   # Large min_group_frac
-  expect_warning(pagfl(y ~ ., data = data, index = c("i", "t"), lambda = 1, verbose = T, min_group_frac = .5))
+  expect_warning(pagfl(y ~ ., data = data, index = c("i", "t"), lambda = 1, verbose = TRUE, min_group_frac = .5))
   # 0 ming_group_frac
-  expect_no_error(pagfl(y ~ ., data = data, index = c("i", "t"), lambda = 1, verbose = T, min_group_frac = NULL))
+  expect_no_error(pagfl(y ~ ., data = data, index = c("i", "t"), lambda = 1, verbose = F, min_group_frac = NULL))
   # Negative ming_group_frac
-  expect_error(pagfl(y ~ ., data = data, index = c("i", "t"), lambda = 1, verbose = T, min_group_frac = -.5))
+  expect_error(pagfl(y ~ ., data = data, index = c("i", "t"), lambda = 1, verbose = F, min_group_frac = -.5))
   # Force some trivial groups
   sim_smallNk <- readRDS(test_path("fixtures", "pagfl_pls_sim_smallNk.rds"))
-  expect_no_error(pagfl(y ~ ., data = sim_smallNk$data, n_periods = 75, lambda = 2, min_group_frac = .3))
+  expect_no_error(pagfl(y ~ ., data = sim_smallNk$data, n_periods = 75, lambda = 2, min_group_frac = .3, verbose = F))
   # Intercept
   data_star_2 <- data
   data_star_2$c <- 1
-  expect_error(pagfl(y ~ c, data = data_star_2, index = c("i", "t"), lambda = 1, verbose = T))
+  expect_error(pagfl(y ~ c, data = data_star_2, index = c("i", "t"), lambda = 1, verbose = F))
   # Force only one group
-  expect_no_error(pagfl(y ~ ., data = data, index = c("i", "t"), lambda = 1e4))
+  expect_no_error(pagfl(y ~ ., data = data, index = c("i", "t"), lambda = 1e6, verbose = F))
 })
 
 test_that("Unbalanced panel pagfl", {
@@ -89,35 +89,35 @@ test_that("grouped_plm inputs", {
   expect_error(grouped_plm(y ~ 1, data = data, groups = groups_0_star, n_periods = 150, verbose = F))
 })
 
-test_that("tv_pagfl inputs", {
+test_that("fuse_time inputs", {
   skip_on_cran()
-  sim <- readRDS(test_path("fixtures", "tv_pagfl_sim.rds"))
+  sim <- readRDS(test_path("fixtures", "fuse_time_sim.rds"))
   y <- sim$y
   data <- as.data.frame(cbind(y = c(y)))
   data$i <- as.character(rep(1:20, each = 100))
   data$t <- rep(1:100, 20)
   # Wrong number of time periods
-  expect_error(tv_pagfl(y ~ 1, data = data, n_periods = 101, lambda = 1, verbose = F))
+  expect_error(fuse_time(y ~ 1, data = data, n_periods = 101, lambda = 1, verbose = F))
   # Char matrix for y
   data_star <- data.frame(y = as.character(c(y)))
-  expect_error(tv_pagfl(y ~ 1, data = data_star, n_periods = 100, lambda = 1, verbose = F))
+  expect_error(fuse_time(y ~ 1, data = data_star, n_periods = 100, lambda = 1, verbose = F))
   # Wrong index variables
   data$a <- stats::rnorm(length(y))
-  expect_error(tv_pagfl(y ~ 1 + a, data = data, index = c("a", "t"), lambda = 1, verbose = F))
-  expect_error(tv_pagfl(y ~ 1, data = data, index = c("c", "t"), lambda = 1, verbose = F))
+  expect_error(fuse_time(y ~ 1 + a, data = data, index = c("a", "t"), lambda = 1, verbose = F))
+  expect_error(fuse_time(y ~ 1, data = data, index = c("c", "t"), lambda = 1, verbose = F))
   # Nonexistent regressor
-  expect_error(tv_pagfl(y ~ 1 + b, data = data, index = c("i", "t"), lambda = 1, verbose = F))
+  expect_error(fuse_time(y ~ 1 + b, data = data, index = c("i", "t"), lambda = 1, verbose = F))
   # Wrong argument
-  expect_error(tv_pagfl(y ~ 1, data = data, index = c("i", "t"), lambda = 1, verbose = F, d = -1))
+  expect_error(fuse_time(y ~ 1, data = data, index = c("i", "t"), lambda = 1, verbose = F, d = -1))
   # Const_coef not in data
-  expect_error(tv_pagfl(y ~ 1, data = data, index = c("i", "t"), lambda = 1, verbose = F, const_coef = "a"))
+  expect_error(fuse_time(y ~ 1, data = data, index = c("i", "t"), lambda = 1, verbose = F, const_coef = "a"))
   # Force only one group with const coef
-  expect_no_error(tv_pagfl(y ~ 1 + a, data = data, index = c("i", "t"), lambda = 1e4, verbose = F, const_coef = "a"))
+  expect_no_error(fuse_time(y ~ 1 + a, data = data, index = c("i", "t"), lambda = 1e4, verbose = F, const_coef = "a"))
 })
 
 test_that("tv_grouped_plm inputs", {
   skip_on_cran()
-  sim <- readRDS(test_path("fixtures", "tv_pagfl_sim.rds"))
+  sim <- readRDS(test_path("fixtures", "fuse_time_sim.rds"))
   y <- sim$y
   data <- as.data.frame(cbind(y = c(y)))
   groups_0 <- sim$groups