+++ NEWS.md
@@ -1,4 +1,15 @@
-# PAGFL 1.1.3 (development version)
+# PAGFL 1.1.4 (development version)
+
+* Added references to new arXiv submission 
+* Renamed `tv_pagfl` to `fuse_time`
+* Updated documentation
+* Slight cosmetic changes to figures
+* Added icon
+* Observations with NAs of irrelevant variables are no longer removed
+* Extended unit tests
+* Added R >= 3.5.0 dependency
+
+# PAGFL 1.1.3 (stable version)
 
 * Added a progress bar displaying the percentage of convergence to `pagfl` and `tv_pagfl` if `verbose` is selected
 * Slight changes in the formatting of the `print` method for `summary` of `pagfl` and `grouped_plm` objects
@@ -9,7 +20,7 @@
 * Added seeds to all function examples
 * Support for RcppArmadillo 14.4.0
 
-# PAGFL 1.1.2 (stable version)
+# PAGFL 1.1.2
 
 * Fixed backwards compatibility issue with the generic functions `fitted` and `resid`
 * Bugfix when passing index variables and an empty formula `y ~ .` for `tv_pagfl` and `grouped_tv_plm`